+++ src/main/java/com/planit/planit/web/controller/MemberController.java
@@ -7,12 +7,10 @@ import com.planit.planit.common.api.member.status.MemberSuccessStatus;
 import com.planit.planit.auth.jwt.UserPrincipal;
 import com.planit.planit.member.service.MemberService;
 import com.planit.planit.member.service.NotificationService;
-import com.planit.planit.web.dto.auth.login.OAuthLoginDTO;
 
 import com.planit.planit.web.dto.member.MemberInfoResponseDTO;
 import com.planit.planit.web.dto.member.notification.NotificationDTO;
 import com.planit.planit.web.dto.member.term.TermAgreementDTO;
-import jakarta.servlet.http.HttpServletRequest;
 import com.planit.planit.web.dto.member.MemberResponseDTO;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
@@ -39,29 +37,7 @@ public class MemberController {
     private final NotificationService notificationService;
 
 
-    @Operation(summary = "idToken 기반 로그인/회원가입", description = "모바일 앱에서 받은 idToken을 검증하여 로그인 또는 회원가입을 처리합니다.")
-    @PostMapping("/sign-in")
-    public ApiResponse<OAuthLoginDTO.LoginResponse> signIn(@RequestBody OAuthLoginDTO.LoginRequest loginRequest) {
-        OAuthLoginDTO.LoginResponse loginResponse = memberService.signIn(loginRequest);
-        log.info("✅ 로그인 or 회원가입 성공 - id: {}, email: {}, name: {}, isNewMember: {}, 약관 동의여부: {}",
-                loginResponse.getId(), loginResponse.getEmail(), loginResponse.getName(), loginResponse.isNewMember(), loginResponse.isSignUpCompleted());
-        return ApiResponse.onSuccess(MemberSuccessStatus.SIGN_IN_SUCCESS, loginResponse);
-    }
-
-    @Operation(summary = "로그아웃", description = "사용자 로그아웃을 처리하고 토큰을 블랙리스트에 추가합니다.")
-    @SecurityRequirement(name = "accessToken")
-    @PostMapping("/sign-out")
-    public ApiResponse<Void> signOut(@AuthenticationPrincipal UserPrincipal principal, HttpServletRequest request) {
-        String authHeader = request.getHeader("Authorization");
-        if (authHeader != null && authHeader.startsWith("Bearer ")) {
-            String accessToken = authHeader.substring("Bearer ".length());
-            memberService.signOut(principal.getId(), accessToken);
-            log.info("✅ 로그아웃 성공 - id: {}, accessToken: {}...", principal.getId(), accessToken.substring(0, 10));
-        }
-        return ApiResponse.onSuccess(MemberSuccessStatus.SIGN_OUT_SUCCESS, null);
-    }
-
-    @Operation(summary = "약관 동의 완료", description = "사용자가 약관에 동의했음을 저장하고 isSignUpCompleted를 true로 갱신합니다.")
+    @Operation(summary = "[TERM] 약관 동의 완료", description = "사용자가 약관에 동의했음을 저장하고 isSignUpCompleted를 true로 갱신합니다.")
     @SecurityRequirement(name = "accessToken")
     @PostMapping("/terms")
     public ApiResponse<Void> agreeTerms(
@@ -72,7 +48,7 @@ public class MemberController {
         return ApiResponse.onSuccess(MemberSuccessStatus.TERM_AGREEMENT_COMPLETED, null);
     }
 
-    @Operation(summary = "모든 약관 URL 조회", description = "최신 약관 HTML 파일들의 URL과 버전을 반환합니다.")
+    @Operation(summary = "[TERM] 모든 약관 URL 조회", description = "최신 약관 HTML 파일들의 URL과 버전을 반환합니다.")
     @GetMapping("/terms")
     public ApiResponse<Map<String, Map<String, String>>> getTermsUrls() {
         Map<String, Map<String, String>> termsInfo = termService.getAllTermsUrls();
