+++ src/main/java/com/planit/planit/web/dto/auth/OAuthLoginDTO.java
@@ -1,4 +1,4 @@
-package com.planit.planit.web.dto.auth.login;
+package com.planit.planit.web.dto.auth;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import io.swagger.v3.oas.annotations.media.Schema;
