+++ src/test/java/com/planit/planit/auth/fixture/FakeOAuth2User.java
@@ -1,4 +1,4 @@
-package com.planit.planit.auth;
+package com.planit.planit.auth.fixture;
 
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.oauth2.core.user.OAuth2User;
