+++ src/test/java/com/planit/planit/member/repository/MemberRepositoryTest.java
@@ -1,8 +1,8 @@
-package com.planit.planit.member;
+package com.planit.planit.member.repository;
 
+import com.planit.planit.member.Member;
 import com.planit.planit.member.enums.Role;
 import com.planit.planit.member.enums.SignType;
-import com.planit.planit.member.repository.MemberRepository;
 import org.junit.jupiter.api.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
